Added admin methods

+++ src/main/java/ru/service/jchat/services/MessageService.java
@@ -9,6 +9,7 @@ import ru.service.jchat.models.request.MessageCreateRequest;
 import ru.service.jchat.models.request.MessagePinRequest;
 import ru.service.jchat.models.request.MessageUpdateRequest;
 import ru.service.jchat.models.response.dto.MessageDTO;
+import ru.service.jchat.repositories.ChatRepository;
 import ru.service.jchat.repositories.MessageRepository;
 import ru.service.jchat.repositories.UserRepository;
 import ru.service.jchat.services.transfer.MessageTransferService;
@@ -16,6 +17,7 @@ import ru.service.jchat.services.transfer.MessageTransferService;
 import javax.persistence.EntityNotFoundException;
 import java.time.ZonedDateTime;
 import java.util.Map;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -23,11 +25,13 @@ public class MessageService {
     private final MessageRepository messageRepository;
     private final MessageTransferService messageTransferService;
     private final UserRepository userRepository;
+    private final ChatRepository chatRepository;
 
-    public MessageService(MessageRepository messageRepository, MessageTransferService messageTransferService, UserRepository userRepository) {
+    public MessageService(MessageRepository messageRepository, MessageTransferService messageTransferService, UserRepository userRepository, ChatRepository chatRepository) {
         this.messageRepository = messageRepository;
         this.messageTransferService = messageTransferService;
         this.userRepository = userRepository;
+        this.chatRepository = chatRepository;
     }
 
     public MessageDTO getById(Long id) {
@@ -47,8 +51,13 @@ public class MessageService {
         return messageTransferService.messageToDto(messageRepository.save(message));
     }
 
-    public MessageDTO update(Long id, MessageUpdateRequest request) {
+    public MessageDTO update(Long id, MessageUpdateRequest request, JwtAuthentication authInfo) throws Exception {
         MessageEntity message = messageRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Message with id " + id + " not found"));
+
+        if (!Objects.equals(message.getUser().getEmail(), authInfo.getPrincipal())) {
+            throw new Exception("You can't edit other people's messages");
+        }
+
         message.setText(request.getText());
         message.setPinned(request.getPinned());
         message.setUpdated(ZonedDateTime.now());
@@ -56,7 +65,13 @@ public class MessageService {
         return messageTransferService.messageToDto(messageRepository.save(message));
     }
 
-    public void delete(Long id) {
+    public void delete(Long id, JwtAuthentication authInfo) throws Exception {
+        MessageEntity message = messageRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Message with id " + id + " not found"));
+
+        if (!Objects.equals(message.getUser().getEmail(), authInfo.getPrincipal())) {
+            throw new Exception("You can't delete other people's messages");
+        }
+
         messageRepository.deleteById(id);
     }
 
@@ -81,5 +96,15 @@ public class MessageService {
         messageRepository.save(message);
 
         return messageTransferService.messageToDto(message);
+    }
+
+    public void deleteMessageFromChat(Long messageId, Long chatId) throws Exception {
+        MessageEntity message = messageRepository.findById(messageId).orElseThrow(() -> new EntityNotFoundException("Message with id " + messageId + " not found!"));
+
+        if (!Objects.equals(message.getChat().getId(), chatId)) {
+            throw new Exception("Message is not in the chat");
+        }
+
+        messageRepository.deleteById(messageId);
     }
 }
